requirements.txt
python-dotenv
requests
aiogram
aiohttp 
supabase